+++ TP/Nacional.java
@@ -3,7 +3,7 @@ import java.util.Date;
 public class Nacional extends Mudanza {
     private int canObjMax;
 
-    public Nacional(String origen, String destino, int km, int precio, Date horario, String tipoTransporte,
+    public Nacional(String origen, String destino, int km, double precio, Date horario, String tipoTransporte,
             int canObjMax) {
         super(origen, destino, km, precio, horario, tipoTransporte);
         this.canObjMax = canObjMax;
